separate group and remove set
