+++ pyiron_atomistics/sphinx/base.py
@@ -41,7 +41,7 @@ from pyiron_atomistics.sphinx.output_parser import (
     SphinxLogParser,
 )
 from pyiron_atomistics.sphinx.input_writer import (
-    write_spin_constraints, copy_potentials
+    write_spin_constraints, copy_potentials, Group
 )
 from pyiron_atomistics.sphinx.util import sxversions
 from pyiron_atomistics.sphinx.volumetric_data import SphinxVolumetricData
@@ -1964,95 +1964,6 @@ def get_structure_group(structure, use_symmetry=True, keep_angstrom=False):
             {"operator": {"S": "[[1,0,0],[0,1,0],[0,0,1]]"}}
         )
     return structure_group
-
-
-class Group(DataContainer):
-    """
-    Dictionary-like object to store SPHInX inputs.
-
-    Attributes (sub-groups, parameters, & flags) can be set
-    and accessed via dot notation, or as standard dictionary
-    key/values.
-
-    `to_{job_type}` converts the Group to the format
-    expected by the given DFT code in its input files.
-    """
-
-    def set(self, name, content):
-        self[name] = content
-
-    def set_group(self, name, content=None):
-        """
-        Set a new group in SPHInX input.
-
-        Args:
-            name (str): name of the group
-            content: content to append
-
-        This creates an input group of the type `name { content }`.
-        """
-        if content is None:
-            self.create_group(name)
-        else:
-            self.set(name, content)
-
-    def set_flag(self, flag, val=True):
-        """
-        Set a new flag in SPHInX input.
-
-        Args:
-            flag (str): name of the flag
-            val (bool): boolean value
-
-        This creates an input flag of the type `name = val`.
-        """
-        self.set(flag, val)
-
-    def set_parameter(self, parameter, val):
-        """
-        Set a new parameter in SPHInX input.
-
-        Args:
-            parameter (str): name of the flag
-            val (float): parameter value
-
-        This creates an input parameter of the type `parameter = val`.
-        """
-        self.set(parameter, val)
-
-    def remove(self, name):
-        if name in self.keys():
-            del self[name]
-
-    def to_sphinx(self, content="__self__", indent=0):
-        if content == "__self__":
-            content = self
-
-        def format_value(v):
-            if isinstance(v, bool):
-                if v:
-                    return ";"
-                else:
-                    return " = false;"
-            elif isinstance(v, Group):
-                if len(v) == 0:
-                    return " {}"
-                else:
-                    return " {\n" + self.to_sphinx(v, indent + 1) + indent * "\t" + "}"
-            else:
-                if isinstance(v, np.ndarray):
-                    v = v.tolist()
-                return " = {!s};".format(v)
-
-        line = ""
-        for k, v in content.items():
-            if isinstance(v, Group) and len(v) > 0 and not v.has_keys():
-                for vv in v.values():
-                    line += indent * "\t" + str(k) + format_value(vv) + "\n"
-            else:
-                line += indent * "\t" + str(k) + format_value(v) + "\n"
-
-        return line
 
 
 class Output:
